response module
response.successful function
Monkey-patching requests.Response.successful
